Add the ability to create a define statement
I did not update the readme, but this allows for the 'define' statements that are required for reasoning in virtuoso. This is based on:
http://vos.openlinksw.com/owiki/wiki/VOS/VirtSPARQLReasoningTutorial#Step%206.%3A%20SPARQL%20Inference%20Queries

I utilized your basic structure and have tested the following:
setDefaultDefines({...<defines object>...});
addDefines({...<defines object>...});
setOptions(<encoding type>, <prefixes object>, <IRI>, {...<defines object>...});

The defines object looks just like the prefixes object. From the example above this would be:
{
   "input:inference":  "urn:owl:inference:rules:tests"
}